Issue #3: local cluster.slurm.cheaha.json being ignored
Changes to ./cluster.slurm.cheaha.json seemed to have no effect.

When run as "snakemakeslurm --debug" it showed only the global cluster json file being put on the snakemake command line.

Problem turned out to be a typo. It was finding the local json file, but then recording it a mis-spelled variable.

I corrected the spelling error, refactored the json search code into a loop, and had it print out (in debug mode) when it finds a local file and uses it.